Added ResponseStatusException

+++ src/main/java/com/deloitte/deloitter/user/service/UserServiceImpl.java
@@ -4,6 +4,10 @@ import com.deloitte.deloitter.user.entity.User;
 import com.deloitte.deloitter.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 public class UserServiceImpl implements IUserService{
@@ -20,8 +24,9 @@ public class UserServiceImpl implements IUserService{
     public User saveUser(User user){
         if (userRepository.findUserByName(user.getName()) == null) {
             return userRepository.save(user);
+        } else {
+            throw new ResponseStatusException(NOT_FOUND, "Unable to find user with this name");
         }
-        return null;
     }
 
     @Override
@@ -31,12 +36,18 @@ public class UserServiceImpl implements IUserService{
             upUser.setName(user.getName());
             upUser.setPassword(user.getPassword());
             return userRepository.save(upUser);
+        } else {
+            throw new ResponseStatusException(NOT_FOUND, "Unable to find user with this name");
         }
-        return null;
     }
 
     @Override
-    public void deleteUserByName(String name){
-        userRepository.delete(findUserByName(name));
+    public String deleteUserByName(String name){
+        if(userRepository.findUserByName(name) != null){
+            userRepository.delete(findUserByName(name));
+            return "User deleted";
+        } else {
+            throw new ResponseStatusException(NOT_FOUND, "Unable to find user with this name");
+        }
     }
 }
